+++ tests/testthat/test-python-numpy.R
@@ -71,4 +71,22 @@ test_that("np_array can reshape existing numpy arrays", {
   expect_equal(py_to_r(a$shape), list(2L, 2L, 2L))
 })
 
+test_that("numpy dim and length functions work", {
+  skip_if_no_numpy()
+  
+  test_array <- function(a) {
+    expect_equal(dim(a), c(2,4))
+    dim(a) <- c(2,2,2)
+    expect_equal(dim(a), c(2,2,2))
+    expect_equal(length(a), 8)
+  }
+  
+  # test no-convert numpy array
+  a <- np_array(c(1L:8L), dim = c(2,4), dtype = "float32")
+  test_array(a)
+  
+  # test convertable numpy array
+  np <- import("numpy")
+  test_array(r_to_py(matrix(c(1:8), nrow = 2, ncol = 4)))
+})